+++ examples/bzlmod/__main__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __init__ import main
+from lib import main
 
 if __name__ == "__main__":
     print(main([["A", 1], ["B", 2]]))
