+++ examples/bzlmod/lib.py
